refactor(ci-hygiene): split badge generation command (#0000)
