+++ crates/perl-ci-hygiene/src/commands/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod badges;
 pub(crate) mod doc_paths;
